requirements.txt
Pillow==11.0.0tkcalendar==1.6.1